ISPN-961 - Implement a provider for Spring's Cache Abstraction
Up to now, we had to set "forkMode" to "perTest" in
maven-surefire-plugin's configuration. Otherwise, an
EmbeddedCacheManager created by a test case that was being left in a
running state by that test case would cause creation of another
EmbeddedCacheManager to fail. This was due to the fact that
EmbeddedCacheManager instances are cached, and each no two
EmbeddedCacheManager instances may share the same JMX domain.

Setting "forkMode" to "perTest", however, is ignored when running the
maven-surefire-plugin for creating a cobertura report. This seems to
be a limitation in the maven-cobertura-plugin. Consequently, the
generated site would report bogus test failures.

Therefore:

 * Make each test case stop created CacheManager instances.
 * For performance reasons, switch "forkMode" back to "once".